Refactor, rearrange: Message, Header, as_enum at top level; other contents of messages.py moved to new messages package
