+++ src/gwproto/__init__.py
@@ -1,36 +1,40 @@
 import gwproto.enums as enums
 import gwproto.messages as messages
 import gwproto.property_format as property_format
-
-from .decoders import Decoder
-from .decoders import DecoderItem
-from .decoders import Decoders
-from .decoders_factory import DecoderExtractor
-from .decoders_factory import MessageDiscriminator
-from .decoders_factory import OneDecoderExtractor
-from .decoders_factory import PydanticExtractor
-from .decoders_factory import create_message_payload_discriminator
-from .decoders_factory import get_pydantic_literal_type_name
-from .decoders_factory import gridworks_message_decoder
-from .decoders_factory import pydantic_named_types
-from .errors import MpSchemaError
+from gwproto.decoders import Decoder
+from gwproto.decoders import DecoderItem
+from gwproto.decoders import Decoders
+from gwproto.decoders_factory import DecoderExtractor
+from gwproto.decoders_factory import MessageDiscriminator
+from gwproto.decoders_factory import OneDecoderExtractor
+from gwproto.decoders_factory import PydanticExtractor
+from gwproto.decoders_factory import create_message_payload_discriminator
+from gwproto.decoders_factory import get_pydantic_literal_type_name
+from gwproto.decoders_factory import gridworks_message_decoder
+from gwproto.decoders_factory import pydantic_named_types
+from gwproto.errors import MpSchemaError
+from gwproto.message import Header
+from gwproto.message import Message
+from gwproto.message import as_enum
 
 
 __all__ = [
-    # top level
+    "as_enum",
+    "create_message_payload_discriminator",
     "Decoder",
     "DecoderItem",
     "Decoders",
-    "create_message_payload_discriminator",
     "DecoderExtractor",
-    "gridworks_message_decoder",
-    "MessageDiscriminator",
-    "OneDecoderExtractor",
+    "enums",
     "get_pydantic_literal_type_name",
+    "gridworks_message_decoder",
+    "Header",
+    "Message",
+    "messages",
+    "MessageDiscriminator",
+    "MpSchemaError",
+    "OneDecoderExtractor",
+    "property_format",
     "pydantic_named_types",
     "PydanticExtractor",
-    "enums",
-    "MpSchemaError",
-    "property_format",
-    "messages",
 ]
